pyproject.toml
[tool.poetry]
name = "arena-api"
version = "1.0.0"
description = "Arena project"
authors = ["amadou sy <amsada93@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.9"
gunicorn = "^20.1.0"
django-environ = "^0.8.1"
psycopg2-binary = "^2.9.3"
graphene-django = "^2.15.0"
django-filter = "^21.1"
django-graphql-jwt = "^0.3.4"
graphene-graphiql-explorer = "^0.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
factory-boy = "^3.2.1"
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
